+++ backend/app.js
@@ -1,36 +1,36 @@
-var createError = require('http-errors');
-var express = require('express');
-var path = require('path');
-var cookieParser = require('cookie-parser');
-var logger = require('morgan');
+var createError = require("http-errors");
+var express = require("express");
+var path = require("path");
+var cookieParser = require("cookie-parser");
+var logger = require("morgan");
 
-
-var indexRouter = require('./routes/index');
-var usersRouter = require('./routes/users');
-var commentsRouter = require('./routes/comments');
-var boardRouter = require('./routes/board');
-var posting = require('./routes/posting');
-var articleRouter = require('./routes/article');
+var indexRouter = require("./routes/index");
+var adminRouter = require("./routes/admin");
+var usersRouter = require("./routes/users");
+var commentsRouter = require("./routes/comments");
+var boardRouter = require("./routes/board");
+var postRouter = require("./routes/post");
+var articleRouter = require("./routes/article");
 
 var app = express();
 
 // view engine setup
-app.set('views', path.join(__dirname, 'views'));
-app.set('view engine', 'pug');
+app.set("views", path.join(__dirname, "views"));
+app.set("view engine", "pug");
 
-app.use(logger('dev'));
+app.use(logger("dev"));
 app.use(express.json());
 app.use(express.urlencoded({ extended: false }));
 app.use(cookieParser());
-app.use(express.static(path.join(__dirname, 'public')));
+app.use(express.static(path.join(__dirname, "public")));
 
-
-app.use('/', indexRouter);
-app.use('/api/users', usersRouter);
-app.use('/api/comments', commentsRouter);
-app.use('/api/board', boardRouter);
-app.use('/api/post', posting);
-app.use('/api/article', articleRouter);
+app.use("/", indexRouter);
+app.use("/api/admin", adminRouter);
+app.use("/api/users", usersRouter);
+app.use("/api/comments", commentsRouter);
+app.use("/api/board", boardRouter);
+app.use("/api/post", postRouter);
+app.use("/api/article", articleRouter);
 
 /* 문제점 ?????
 app.use(function(req, res, next) {
@@ -50,19 +50,19 @@ app.options('/api/users/signUp', (req, res) => {
 */
 
 // catch 404 and forward to error handler
-app.use(function(req, res, next) {
-    next(createError(404));
+app.use(function (req, res, next) {
+  next(createError(404));
 });
 
 // error handler
-app.use(function(err, req, res, next) {
-    // set locals, only providing error in development
-    res.locals.message = err.message;
-    res.locals.error = req.app.get('env') === 'development' ? err : {};
+app.use(function (err, req, res, next) {
+  // set locals, only providing error in development
+  res.locals.message = err.message;
+  res.locals.error = req.app.get("env") === "development" ? err : {};
 
-    // render the error page
-    res.status(err.status || 500);
-    res.render('error');
+  // render the error page
+  res.status(err.status || 500);
+  res.render("error");
 });
 
 module.exports = app;